Question3/requirements.txt
tenacity
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
snowflake-connector-python
pyspark
requests
apache-airflow
python-dotenv
tenacity